// lib/types.ts
export type ProfileType = {
  id: number
  username: string
  email: string
  name: string
  slug: string
  nickname: string
  gender: string
  image: string
  dob: string
  website: string
  contact: string
  contact_email: string
  linkedin: string
  github: string
  address: string
  about: string
  is_portfolio_user: string
  resume_link: string
  is_active: string
  is_staff: string
  is_superuser: string
  date_joined: string
  last_login: string
  updated_at: string
}